+++ lib/commands/mailbox_storage.ml
@@ -23,7 +23,7 @@ struct
   type t = string
 
   (* user *)
-  let create config user mailbox =
+  let create config user mailbox keys =
     return user
 
   let exists t =